move order notifier to providers directory
